drag cassse

+++ app/Observers/H5pObserver.php
@@ -37,9 +37,17 @@ class H5pObserver
     public function updated(h5pLesson $h5pLesson)
     {
         if($h5pLesson->isDirty('lesson_id')){
+
             $lesson = Lesson::find($h5pLesson->lesson_id);
             $course_id = $lesson->courseSegment->course_id;
-            UserSeen::where('lesson_id',$h5pLesson->getOriginal('lesson_id'))->where('item_id',$h5pLesson->content_id)->where('type','h5p')->delete();
+            $class_id = $lesson->courseSegment->segmentClasses[0]->classLevel[0]->class_id;
+
+            $old_lesson = Lesson::find($h5pLesson->getOriginal('lesson_id'));
+            $old_class_id = $old_lesson->courseSegment->segmentClasses[0]->classLevel[0]->class_id;
+            
+            if($old_class_id != $class_id)
+                UserSeen::where('lesson_id',$h5pLesson->getOriginal('lesson_id'))->where('item_id',$h5pLesson->content_id)->where('type','h5p')->delete();
+
             $this->report->calculate_course_progress($course_id);
         }
     }
